revamp the data hook and usages
