+++ meal-ui/src/pages/Recipes/ManageRecipesTable.jsx
@@ -1,39 +1,10 @@
-import { useState } from 'react';
 import { Grid, Column } from '@carbon/react';
 import { RecipeGrid } from './RecipeGrid';
 import { Actions } from './Actions';
+import useManagementTable from './useManagementTable';
 
-export function ManageRecipesTable({ recipes, refreshGrid, archiveRecipes }) {
-	const [selectedItems, setSelectedItems] = useState([]);
-
-	const addSelectedItem = newItem => {
-		setSelectedItems([...selectedItems, newItem]);
-	};
-
-	const removeSelectedItem = oldItem => {
-		const idx = selectedItems.indexOf(oldItem);
-
-		if (idx > -1) {
-			const tempItems = [...selectedItems];
-			tempItems.splice(idx, 1);
-			setSelectedItems(tempItems);
-		} else {
-			console.warn('cannot remove unknown item from grid');
-		}
-	};
-
-	const addOrRemoveSelectedItem = (item, toggle) => {
-		toggle ? addSelectedItem(item) : removeSelectedItem(item);
-	};
-
-	const onArchive = async () => {
-		await archiveRecipes(selectedItems);
-		refreshGrid();
-	};
-
-	const onViewArchived = event => {
-		refreshGrid({ includeArchived: event.target.checked });
-	};
+export function ManageRecipesTable() {
+	const { recipes, onAddOrRemove, onArchive, onViewArchived } = useManagementTable();
 
 	return (
 		<Grid>
@@ -46,7 +17,7 @@ export function ManageRecipesTable({ recipes, refreshGrid, archiveRecipes }) {
 					<h2>Manage Recipes</h2>
 				</div>
 				<RecipeGrid
-					addOrRemoveSelectedItem={addOrRemoveSelectedItem}
+					addOrRemoveSelectedItem={onAddOrRemove}
 					recipes={recipes}
 				/>
 				<Actions
